fix(review): recognize successful check-verdict sessions
Match check-verdict lookups to the requested review scope and accept successful reviewer_sub_session records when they are the latest matching verdict. Split review_cmd_check_verdict tests into a dedicated file to satisfy the monolith gate, add coverage for range and reviewer_sub_session lookups, and include the required workspace patch version bump triggered by pre-commit.
